ema calculation logic untested

+++ src/main/java/com/epex/cryptonius/repository/TickerEntity.java
@@ -3,6 +3,8 @@ package com.epex.cryptonius.repository;
 import lombok.Data;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Data
 @Document(collection = "ticker_db")
 public class TickerEntity {
@@ -15,17 +17,17 @@ public class TickerEntity {
     private String base_unit;
     private String quote_unit;
 
-    private Double low;     //24 hrs lowest price of base asset
-    private Double high;    //24 hrs highest price of base asset
-    private Double last;    //Last traded price in current market
+    private BigDecimal low;     //24 hrs lowest price of base asset
+    private BigDecimal high;    //24 hrs highest price of base asset
+    private BigDecimal last;    //Last traded price in current market
 
     private String type;
 
-    private Double open;    //Market Open price 24hrs ago
-    private Double volume;  //Last 24hrs traded volume
+    private BigDecimal open;    //Market Open price 24hrs ago
+    private BigDecimal volume;  //Last 24hrs traded volume
 
-    private Double sell;    //Top ask order price
-    private Double buy;     //Top bid order price
+    private BigDecimal sell;    //Top ask order price
+    private BigDecimal buy;     //Top bid order price
 
     private Long at;
